Remove debit/credit transactions, and validations
The single table inheritence model of the debit and credit transactions
added more complexity and overhead than it was worth in code. I've
removed it in favor of a positive or negative point balance.

The range validations on point transactions are also being removed as
there are many common cases of admins wishing to manually grant points
outside of the acceptable ranges. This places the burden of validations
on point usage on the redemption mechanism, but is a good choice
overall.